requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.13
django-cors-headers==3.13.0
django-env==0.2.1
django-heroku==0.3.1
django-phonenumber-field==6.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
phonenumbers==8.12.50
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.19.2
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
